WIP Just clean up and try to check an anomaly
Appears a rename done in Visual Studio (which does not register as a code/syntax change, and does not mark file as UNSaved) throws Git into a bit of a confused state, At least that is what seemed to have happened as I got a Git warning in Unity Console after I did such a rename.

Just not sure as I am working on and constantly changing animations which may cause erratic behaviors just by itself.