WIP Cube Game
Solved cube placement issues with one minor cosmetic issue remaining (spurious miss of a finger Up event). Not tested yet on hand-held so it may or may not even carry over to target devices... Gameplay remains correct so we'll move on for now.
On to actually making the 4uku game puzzle(s).